Update Tradutor_Legenda_GPU.py
Adicionado ao código para percorrer todas as subpastas em que o arquivo `Tradutor_Legenda_GPU.py` estiver, ele irá em busca de arquivos `.srt` e salvará as traduções na mesma subpasta dos arquivos originais, adicionada também a verificação para que o código pule arquivos que já foram traduzidos.

+++ Tradutor_Legenda_GPU.py
@@ -32,24 +32,32 @@ def translate_srt(file_path):
 def main():
     print("Iniciando o processo de tradução...")
     current_dir = os.path.dirname(os.path.abspath(__file__))
-
-    for file_name in os.listdir(current_dir):
-        if file_name.endswith('_en.srt'):
-            file_path = os.path.join(current_dir, file_name)
-            print(f"Translating file: {file_path}")
-            start_time = time.time()
-            try:
-                translated_lines = translate_srt(file_path)
+    
+    # Usar os.walk para procurar arquivos em subpastas
+    for root, dirs, files in os.walk(current_dir):
+        for file_name in files:
+            if file_name.endswith('_en.srt'):
+                file_path = os.path.join(root, file_name)
                 new_file_name = file_name.replace('_en.srt', '_pt-br.srt')
-                new_file_path = os.path.join(current_dir, new_file_name)
-                with open(new_file_path, 'w', encoding='utf-8') as new_file:
-                    new_file.writelines(translated_lines)
-                end_time = time.time()
-                elapsed_time = end_time - start_time
-                print(f"File saved as: {new_file_path}")
-                print(f"Translation completed in {elapsed_time:.2f} seconds")
-            except Exception as e:
-                print(f"Error translating file {file_path}: {e}")
+                new_file_path = os.path.join(root, new_file_name)
+                
+                # Verificar se o arquivo traduzido já existe
+                if os.path.exists(new_file_path):
+                    print(f"Translation already exists for: {new_file_path}")
+                    continue
+                
+                print(f"Translating file: {file_path}")
+                start_time = time.time()
+                try:
+                    translated_lines = translate_srt(file_path)
+                    with open(new_file_path, 'w', encoding='utf-8') as new_file:
+                        new_file.writelines(translated_lines)
+                    end_time = time.time()
+                    elapsed_time = end_time - start_time
+                    print(f"File saved as: {new_file_path}")
+                    print(f"Translation completed in {elapsed_time:.2f} seconds")
+                except Exception as e:
+                    print(f"Error translating file {file_path}: {e}")
 
 if __name__ == "__main__":
     main()
